Fixes runtime crashes on HashLink
In RenderUtil, HashLink fails to convert renderMode:String and profile:Dynamic to the respective concrete types. Fixed by explicitly assigning the values to local variables with the correct types.

In Painter.hx, HashLink assumes the keys of _stencilReferenceValues:Dictionary<Dynamic, UInt> are actually strings and fails when trying to cast the keys to String. Fixed by directly using ObjectMap.